requirements.txt
aiohttp==3.6.3
antiorm==1.2.1
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.4
chardet==3.0.4
db-sqlite3==0.0.1
db==0.1.1
discord.py==1.5.1
idna==3.1
multidict==4.7.6
pycparser==2.20
python-dotenv==0.15.0
six==1.15.0
yarl==1.5.1
